+++ trunk/vista/src/vista/objeto/BolaBolicheView.h
@@ -12,7 +12,7 @@ class SimpleEditorAnguloFijo;
 namespace CLIENTE {
 class BolaBolicheView : public ObjetoView {
 public:
-	BolaBolicheView(int x, int y, int w, int h, SDL_Texture * textura);
+	BolaBolicheView(float x, float y, float w, float h, SDL_Texture * textura);
 	virtual ~BolaBolicheView();
 	void update(ViewMsj *);
 };
