+++ trunk/vista/src/vista/objeto/MotorView.cpp
@@ -9,11 +9,11 @@
 #include "../../ConstantesVista.h"
 
 namespace CLIENTE {
-//MotorView::MotorView(int x, int y, int w, int h, SDL_Texture * textura,SimpleEditorAnguloFijo * editor): ObjetoView(x, y, w, h,textura, editor) {
+//MotorView::MotorView(float x, float y, float w, float h, SDL_Texture * textura,SimpleEditorAnguloFijo * editor): ObjetoView(x, y, w, h,textura, editor) {
 //
 //}
 
-MotorView::MotorView(int x, int y, int w, int h, SDL_Texture * textura): ObjetoView(x, y, w, h,textura) {
+MotorView::MotorView(float x, float y, float w, float h, SDL_Texture * textura): ObjetoView(x, y, w, h,textura) {
 
 }
