+++ trunk/vista/src/vista/objeto/SogaView.cpp
@@ -8,7 +8,7 @@
 #include "SogaView.h"
 namespace CLIENTE {
 
-SogaView::SogaView(int x, int y, SDL_Texture * textura): UnionView(x, y, textura) {
+SogaView::SogaView(float x, float y, SDL_Texture * textura): UnionView(x, y, textura) {
 }
